refactor: extract `Post`(s) components in separate files
